Remove requests and maptiler, use geonames and zipcodes exlusively
After talking to some users, I came to realize that the
"request" type adds complexity for now real value.

Drop it for now. All entries are offers now.

Also, the map seems to confuse many users as well. The implementation is now
purely based on ZIP codes, and a radius search filter is added.